Overview img
The image shows the scene setup and the main blueprint classes responsible for object interaction, enemy behaviors, and gameplay mechanics, as well as the structured scene layout.            **Main Blueprint classes used in the VR Rehabilitation Game, developed in Unreal Engine 5:

Basketball – Handles ball spawning and interactions.
Enemy – Base class for AI enemies with animations and behaviors.
Robot – Enemy with a health bar, takes damage, and explodes on destruction.
Witch – AI character that reacts to gunshots and moves toward the player.**